+++ src/components/molecules/reservation-time-picker/ReservationTimePicker.tsx
@@ -2,7 +2,7 @@ import { Calendar } from '@/components/ui/calendar';
 import Input from '@/components/atoms/input/Input';
 import Image from 'next/image';
 import useTimeSlot from '@/models/activity/use-time-slot';
-import TimeSlotList from '../activity-registration-form/TimeSlotList';
+import TimeSlotList from './TimeSlotList';
 import { TimeSlotData } from '@/types/activity';
 import TimeSlotInput from '@/components/atoms/input/TimeSlotInput';
 import Select from '../select/Select';
